move compare data to __init__ :D
